Minor alteration to make the docs to accomodate toc
The toc that generates the navigation on the website from these docs
is too lomg with so many options now.
By setting a few items as deeper headings we can hide them in the toc
Probably not a great thing to be editing the docs for what is really
a related application issue, but will do for now.