documentation: reference: JSON schema iframe create override style
The docson widget tries to control its own height properties and uses
window events, this results in additional unnecessary scrollbars in the
container div which aren't needed as they already exist on the iframe.

We change to using the div simply as a way to easily reference the
iframe element. As the id for the iframe contains punctuation
and we may want to use this style elsewhere.