Update codelists.md
Update to field titles to align with updates to schema. For the regrant type codelist, the codelist CSV has been titled regrantType.csv - so I have aligned the documentation to this. This is a change to what was presented in the proposal 'For Regrant Type' but it seems easier than amending the CSV at this point.