Codelists: Country titles for countryCodes.csv
Added country titles to countryCodes.csv to provide context and also to
allow software to pull out the title properly.

In addition to this, the code is now wrapped in "" to make ensure that
it is parsed correctly as a string and doens't contain any codes that
can be misinterpreted as control sequences in software parsers.