codelists: added titles to countryCode.csv
In response to #349 which stipulated a requirement for titles in
countrycodes.

List of titles was taken from the ISO 3166-1 alpha-2 codes, which
corresponds with the codes used in this codelist. The wikipedia article
suggests that these are the titles used by the ISO 3166 maintenance
agency.

It should be acknowledged that this is not necessarily a value-neutral
list of names.

Amendment was done mechanically through copying the wikipedia table to a
spreadsheet, exporting as CSV and then merging via csvkit. Manual
tidying for strings done afterwards