Remove summary-table-schema
Fixes #405

As per the linked issue, the 360-giving-summary-table-schema.json file
does not appear to be used for anything and has not been updated since
2018.

Discussion indicates that it is an artefact from early standard
development (possibly pre flatten-tool days) to define the tabular
serialization.

Tabular serialization is now done directly from the schema via
flatten-tool, and the documentation build pipeline generates this
template from the schema for each new version. This indicates that the
summary table schema is now defunct.

The Governance page of the standard docs does not specify which files
are covered by Governance processes:

* https://standard.threesixtygiving.org/en/latest/about/governance/#versioning-and-upgrade-process

However this file is stated to be normative content by the Normative and
non-normative content policy:

* https://www.360giving.org/wp-content/uploads/360Giving-Standard-Policy-Normative-and-non-normative-content_final_2020-02-28.pdf

I think we are within grounds to remove it for the following reasons:

* it is not the schema which defines the Standard
* further, it directly contradicts the schema (and thus the Standard)
* it is not used for validation or for deriving documentation
* removing this file therefore does not introduce any changes to the
  Standard, normative sections of the docs, or tooling