Schema: Reverted the changes to date validation
Reverted changes to various date validation away from the `format: date`
to use `pattern: $REGEX` as is current practice.

Current thinking is that upgrading the schema to use JSON Schema 2020-12
is a PATCH level change as long as it does not modify any validation
rules.

Current thinking is also that changing the date validation rules, even
to remove the possibility of imaginary dates, constitutes a breaking
change in the validation rules and thus would trigger a MAJOR upgrade;
even if the dates which would be invalidated are in fact imaginary
dates.

Given these two threads; this branch should represent an upgrade to JSON
Schema 2020-12 without the associated change to date validation. This is
more atomic anyway.