toctemplate: updated toctemplate
Updated the toctemplate extension to be compatible with newer sphinx
versions. This extension is required by our theme, so docs builds were
failing without it changing.

For transparency: I used an LLM to help me understand the Sphinx 8.2
docs and the changes, after searching failed me. It suggested code
snippets which I took and adapted, but didn't do a wholesale rewrite