Docs: fix broken links and formatting
The update to use myst-parser rather than commonmark found a bunch of
broken links in the docs. Most of these were simply pointing to the old
locations of various files, which now exist in a different folder
structure.

In a few cases, I needed to add an explicit target to make myst happy