codelists: update currency.csv with titles
Fixes #349

Following the merge of #406 which added titles to countryCode.csv, this
adds titles to currency.csv. There are also some formatting improvements
to the csv file to make parsing a bit more robust.

The schema states that the currency codes are pegged to ISO 4217, which
is the standard for currency codes:

https://en.wikipedia.org/wiki/ISO_4217

The codes included as part of 360Giving's currency.csv list span the
"active codes" and the "historical codes".

In order to preserve the integrity of the existing codelist, I added the
titles by performing a left join onto the existing codelist. This
ensured that no new codes were added, and that only codes with titles
had titles applied to them.

Reviewers should note that some codes share titles e.g. the historical
ISJ and the contemporary ISK are both named "Icelandic króna".