documentation/css: hid rtd-search-form
As part of https://opendataservices.plan.io/issues/58218, while we await
a fix from RTD on the search functionality, we decided that it might be
best to hide the search bar if possible.

This is a basic attempt at that, by including a CSS rule to make the
search form not display.