docs: remove setting of DOCUMENTATION_OPTIONS
Read The Docs suggested that a standout error from their perspective
was:

> Uncaught SyntaxError: redeclaration of const DOCUMENTATION_OPTIONS

And that any customisation we have there might be conflicting.

The only place that this is manually declared is in
_templates/base.html, where this is indeed declared in the Javascript of
the base template.

There appears to be a lot of history associated with the decision, based
on versions of Sphinx.

I've commented out the declaration of this now. There might be more work
to do in _static/documentation_options.js, but this is currently
building on my machine and has correct URLs for the search results.