docs: added rtd redirects.yml file
Added the RTD redirects file which can be used with https://pypi.org/project/readthedocs-cli/ to manage bulk redirects in RTD.

This file is not parsed by the RTD platform and won't affect any
automated builds.

It is useful to have this file in the repo because:

- it provides a machine and human parseable reference for the canonical
  set of redirects we are using on the project

- it can be used to manipulate the redirects more efficiently than in
  the web interface

- given the above, it can be version controlled to provide some
  light-weight governance on the redirects used in the docs. It isn't
  considered normative content because it's not part of the Standard,
  but having it in the repo allows us to use PRs to manage it openly.

- in the future, it can be used as part of the automated build process
  to sync redirects every time it is merged into the main branch.

The file is quite heavily commented to provide inline documentation to
support its use and explain the context of different redirects which are
in place.